+++ CHANGELOG.md
@@ -22,6 +22,9 @@
   * `squareSmall`: ◻
   * `squareSmallFilled`: ◼
 
+* Deprecated `ascii`, instead use `fallback`.
+* Deprecated `glyphOrAscii`, instead use `glyphOrFallback`.
+
 [figures]: https://www.npmjs.com/package/figures
 
 ## 1.0.0